// client/src/pages/Movies/types.tsx
type MoviesTranslationsProps = {
  btnDetails?: string;
  noImage?: string;
  btnApply?: string;
  filtersReleaseDates?: string;
  filtersGenres?: string;
  filterFrom?: string;
  filtersTo?: string;
  errorMessage?: string;
  noData?: string;
  genresErrorMessage?: string;
};

export { MoviesTranslationsProps };